requirements.txt
pyrogram==2.0.39
tgcrypto==1.2.3
python-dotenv==1.0.1
